+++ headers/entities/pawns/Enemy.h
@@ -8,7 +8,7 @@ class Enemy final : public Bot
 {
 	using buuid = boost::uuids::uuid;
 
-	void TickUpdate(float deltaTime) override;
+	void TickUpdate(double deltaTime) override;
 
 public:
 	Enemy(PawnProperty pawnProperty, const std::shared_ptr<BulletPool>& bulletPool, BonusEffectProperty effects = {},
